src/requirements.txt
tqdm==4.55.1
pandas #==1.3.0 on staging in fact
amqp==2.6.1
appdirs==1.4.4
argcomplete==1.10.0
arrow==0.16.0
asgiref==3.4.1
attrs==19.1.0
beautifulsoup4==4.8.0
billiard==3.6.3.0
boto3==1.13.2
botocore==1.16.26
cachetools==2.0.1
celery==4.4.7
certifi==2018.1.18
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
chargebee==2.7.3
distlib==0.3.1
Django==3.2.13
# django-celery-beat==2.0.0
django-celery-beat==1.6.0
django-countries==7.3.2
django-csvimport==2.5
django-debug-toolbar==2.2
django-filter==21.1
# django-jsonfield==1.4.1
django-jwt-auth==0.0.2
django-model-utils==4.2.0
# django-redis==4.12.1
django-redis==4.11.0
# django-storages==1.10.1
django-storages==1.9.1
# django-timezone-field==4.0
django-timezone-field==3.1
django-widget-tweaks==1.4.1
djangorestframework==3.11.2
# djangorestframework-jwt-refresh-token==0.5
https://github.com/s-tatus/django-rest-framework-jwt-refresh-token/tarball/support-django-3.2
docutils==0.15.2
docx2txt==0.8
drf-yasg==1.20.0
EbookLib==0.17.1
extract-msg==0.23.1
factory-boy==2.12.0
Faker==4.0.0
filelock==3.0.12
flake8==3.8.3
forex-python==1.5
freezegun==0.3.13
gis==0.2.1
google-api-core==1.13.0
google-auth==1.4.1
google-cloud-core==1.0.2
google-cloud-translate==1.5.0
google-cloud-vision==0.30.1
googleapis-common-protos==1.5.3
grpcio==1.26.0
gunicorn==20.0.4
identify==1.5.2
idna==2.6
IMAPClient==2.1.0
importlib-metadata==1.4.0
importlib-resources==1.0.2
install==1.3.4
isort==4.3.21
jmespath==0.10.0
kombu==4.6.11
lxml==4.8.0
mccabe==0.6.1
more-itertools==8.1.0
django-multiselectfield==0.1.12
mypy-boto3==1.14.16.0
mypy-boto3-s3==1.14.16.0
newrelic==5.20.0.149
nodeenv==1.5.0
numpy # ==1.21.0 on staging
olefile==0.46
packaging==19.1
pdfminer.six==20191110
pendulum==2.1.0
pg8000==1.13.2
Pillow==7.2.0
pluggy==0.13.1
pre-commit==2.7.1
protobuf==3.8.0
psycopg2-binary==2.9.3
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.6.0
pycountry==20.7.3
pycparser==2.20
pycryptodome==3.9.4
pyflakes==2.2.0
pyheif==0.5.1
PyJWT==1.4.2
pyparsing==2.4.2
pytest==5.3.2
pytest-django==3.8.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-pptx==0.6.18
pytz==2022.1
pytzdata==2019.3
PyYAML==5.3.1
redis==3.5.3
requests==2.18.4
rsa==3.4.2
s3transfer==0.3.3
scramp==1.1.0
simplejson==3.17.2
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
SQLAlchemy==1.3.12
sqlparse==0.3.1
swapper==1.3.0
testing.common.database==2.0.3
testing.postgresql==1.3.0
text-unidecode==1.3
textract==1.6.5
timezonefinder==4.1.0
toml==0.10.1
typing-extensions==3.7.4.2
tzlocal==1.5.1
Unidecode==1.1.1
urllib3==1.22
vine==1.3.0
virtualenv==20.0.31
vulture==2.1
wcwidth==0.1.8
xlrd==1.2.0
XlsxWriter==1.0.2
zipp==1.0.0
django-cors-headers==3.10.1
